Add files via upload
new class 'CoinGame,' which is a mini game that can be accessed when stumbling upon the gnoll and goblin in the armoury. 
NOTE: you need the small pouch of coins in the trunk of confiscated stuff, to have two of three of the following; thespian trait, used soot on yourself, or possess an insignia, and you need to select the option to 'apdopt the unassuming role of another grunt and complain about the weapons...'
Tests are perhaps needed with aftermath on Player.Coins of this interaction and to tie this attribute with the coins in the pouch you possess.